requirements.txt
pytest==2.8.1
pytest-cov==2.1.0
nose==1.3.7
gevent==1.1rc1
flake8==2.5.0
twisted==15.4.0
pycrypto>=2.6.1
requests>=2.4.0
